Add Plugins to Neovim
Add the LanguageClient and echodoc plugins.
LanguageClient allows the integration of the RLS (Rust Language Server)
into vim. Using this it is possible to get features like rename and goto
without having to rely on an IDE or a fully-featured plugin.
The echodoc plugin shows Function Signatures in the command line.