Update zsh config
This updates the `gp` alias to push only the active branch to the
`origin` remote.

The history size is also increased to 1_000_000 lines so I'll never have
to worry about this.

And Ctrl+S to move the cursor right by one word now works without
fucking with preexisting keyboard shortcuts in zsh.