Add variable to disable expensive git PS1 commands
This adds a condition to check for presence of the `GIT_OFF` environment
variable to skip the potentially expensive `git status --porcelain`
command to color the git info red when there are unstaged changes.

Instead of coloring red/green based on HEAD dirtiness, the prompt is
always colored yellow with `GIT_OFF` set. This should clearly indicate
that the git information is currently not retrieved.